requirements.txt
asgiref==3.11.0
Django==5.2.8
Markdown==3.10
PyYAML==6.0.3
sqlparse==0.5.3
